+++ CHANGELOG.rst
@@ -67,6 +67,8 @@ Misc new features:
   releases version 304::
   
     DeprecationWarning: getargs: The 'u' format is deprecated. Use 'U' instead.
+- Added ``timeout`` and ``hard=True/False`` flags to `pysys.process.Process.stop`. Also added logic on Linux which will 
+  automatically attempt a SIGKILL if the SIGTERM times out (though will still raise an exception in this case). 
 
 Fixes: